feat(wallet): refactor WalletView to extend WalletHomeSections for improved structure and maintainability
- Introduced WalletHomeSections class to encapsulate wallet home section interactions.
- Moved relevant methods and properties from WalletView to WalletHomeSections.
- Updated WalletView to extend WalletHomeSections, reducing code duplication.
- Removed unused imports and methods related to scrolling and tapping sections.
- Enhanced interaction methods for better clarity and reusability.

fix(tests): remove unnecessary TestHelpers import and calls in permission system tests

- Eliminated TestHelpers usage in permission system tests to streamline test setup.
- Adjusted test timeouts for consistency and reliability.
