--- tests/helpers.js
@@ -1,568 +1,0 @@
-/* eslint-disable import-x/no-nodejs-modules */
-import { waitFor, web } from 'detox';
-import { getFixturesServerPort } from './framework/fixtures/FixtureUtils';
-import Utilities from './framework/Utilities';
-import { resolveConfig } from 'detox/internals';
-import { createLogger } from './framework/logger';
-import { exec } from 'child_process';
-import { promisify } from 'util';
-
-const logger = createLogger({
-  name: 'TestHelpers',
-});
-const execAsync = promisify(exec);
-
-export default class TestHelpers {
-  /**
-   * @deprecated Use Guestures Class to accomplish this.
-   */
-  static async waitAndTap(elementId, timeout, index) {
-    await waitFor(element(by.id(elementId)))
-      .toBeVisible()
-      .withTimeout(timeout || 8000);
-
-    return element(by.id(elementId))
-      .atIndex(index || 0)
-      .tap();
-  }
-  /**
-   * @deprecated Use Guestures Class to accomplish this.
-   */
-  static async waitAndTapText(text, timeout) {
-    await waitFor(element(by.text(text)))
-      .toBeVisible()
-      .withTimeout(timeout || 8000);
-
-    return element(by.text(text)).tap();
-  }
-  /**
-   * @deprecated Use Guestures Class to accomplish this.
-   */
-  static tap(elementId) {
-    return element(by.id(elementId)).tap();
-  }
-  /**
-   * @deprecated Use Guestures Class to accomplish this.
-   */
-  static tapByDescendentTestID(parentElement, ChildElement) {
-    return element(
-      by.id(parentElement).withDescendant(by.id(ChildElement)),
-    ).tap();
-  }
-  /**
-   * @deprecated Use Guestures Class to accomplish this.
-   */
-  static tapByText(text, index) {
-    return element(by.text(text))
-      .atIndex(index || 0)
-      .tap();
-  }
-  /**
-   * @deprecated Use Guestures Class to accomplish this.
-   */
-  static doubleTapByText(text, index) {
-    return element(by.text(text))
-      .atIndex(index || 0)
-      .multiTap(2);
-  }
-  /**
-   * @deprecated Use Guestures Class to accomplish this.
-   */
-  static tapAtPoint(elementId, point) {
-    return element(by.id(elementId)).tap(point);
-  }
-  /**
-   * @deprecated Use Guestures Class to accomplish this.
-   */
-  static tapItemAtIndex(elementID, index) {
-    return element(by.id(elementID))
-      .atIndex(index || 0)
-      .tap();
-  }
-  /**
-   * @deprecated Use Guestures Class to accomplish this.
-   */
-  static tapItemAtIndexByLabel(elementID, index) {
-    return element(by.label(elementID, index))
-      .atIndex(index || 0)
-      .tap();
-  }
-  /**
-   * @deprecated Use Guestures Class to accomplish this.
-   */
-  static async typeText(elementId, text) {
-    await TestHelpers.tap(elementId);
-    return element(by.id(elementId)).typeText(text);
-  }
-  /**
-   * @deprecated Use Guestures Class to accomplish this.
-   */
-  static async typeNumbers(elementId, text, submitLabel) {
-    await element(by.id(elementId)).replaceText(text.replace('\n', ''));
-    return element(by.label(submitLabel)).atIndex(0).tap();
-  }
-  /**
-   * @deprecated Use Guestures Class to accomplish this.
-   */
-  static async typeTextAndHideKeyboard(elementId, text) {
-    if (device.getPlatform() === 'android') {
-      await TestHelpers.clearField(elementId);
-    }
-    await TestHelpers.typeText(elementId, text + '\n');
-  }
-  /**
-   * @deprecated Use Guestures Class to accomplish this.
-   */
-  static async clearField(elementId) {
-    return element(by.id(elementId)).replaceText('');
-  }
-
-  /**
-   * @deprecated Use Guestures Class to accomplish this.
-   */
-  static async tapAndLongPress(elementId) {
-    await TestHelpers.tap(elementId);
-    return element(by.id(elementId)).longPress(2000);
-  }
-
-  /**
-   * @deprecated Use Guestures Class to accomplish this.
-   */
-  static async tapAndLongPressAtIndex(elementId, index) {
-    return element(by.id(elementId))
-      .atIndex(index || 0)
-      .longPress(2000);
-  }
-
-  /**
-   * @deprecated Use Guestures Class to accomplish this.
-   */
-  static async replaceTextInField(elementId, text) {
-    return element(by.id(elementId)).replaceText(text);
-  }
-  /**
-   * @deprecated Use Guestures Class to accomplish this.
-   */
-
-  static tapAlertWithButton(text, index) {
-    if (device.getPlatform() === 'android') {
-      return element(by.text(text))
-        .atIndex(index || 0)
-        .tap();
-    }
-
-    return element(by.label(text)).atIndex(0).tap();
-  }
-
-  /**
-   * @deprecated Use Guestures Class to accomplish this.
-   */
-  static async waitAndTapByLabel(text, timeout, index) {
-    await waitFor(element(by.label(text)))
-      .toBeVisible()
-      .withTimeout(timeout || 15000);
-
-    return element(by.label(text))
-      .atIndex(index || 0)
-      .tap();
-  }
-  /**
-   * @deprecated Use Guestures Class to accomplish this.
-   */
-  static async tapWebviewElement(elementId) {
-    // this method only words on android: https://wix.github.io/Detox/docs/api/webviews/
-    return web.element(by.web.id(elementId)).tap();
-  }
-  /**
-   * @deprecated Use Guestures Class to accomplish this.
-   */
-  static async swipe(elementId, direction, speed, percentage, xStart, yStart) {
-    await element(by.id(elementId)).swipe(
-      direction,
-      speed,
-      percentage,
-      xStart,
-      yStart,
-    );
-  }
-  /**
-   * @deprecated Use Guestures Class to accomplish this.
-   */
-  static async swipeByLabel(elementId, direction, speed, percentage) {
-    await element(by.label(elementId)).swipe(direction, speed, percentage);
-  }
-
-  /**
-   * @deprecated Use Guestures Class to accomplish this.
-   */
-  static async swipeByText(text, direction, speed, percentage) {
-    await element(by.text(text)).atIndex(0).swipe(direction, speed, percentage);
-  }
-
-  /**
-   * @deprecated Use Guestures Class to accomplish this.
-   */
-  static async scrollTo(scrollViewId, edge) {
-    await element(by.id(scrollViewId)).scrollTo(edge);
-  }
-
-  /**
-   * @deprecated Use Guestures Class to accomplish this.
-   */
-  static async scrollUpTo(elementId, distance, direction) {
-    await element(by.id(elementId)).scroll(distance, direction);
-  }
-
-  static async openDeepLink(inputURL) {
-    await device.launchApp({
-      newInstance: true,
-      url: inputURL,
-      sourceApp: 'io.metamask',
-      launchArgs: {
-        fixtureServerPort: `${getFixturesServerPort()}`,
-        detoxURLBlacklistRegex: Utilities.BlacklistURLs,
-      },
-    });
-  }
-
-  /**
-   * @deprecated Use Assertion Class to accomplish this.
-   */
-  static async checkIfVisible(elementId) {
-    return await waitFor(element(by.id(elementId)))
-      .toBeVisible()
-      .withTimeout(15000);
-  }
-
-  /**
-   * @deprecated Use Assertion Class to accomplish this.
-   */
-  static async checkIfNotVisible(elementId) {
-    return await waitFor(element(by.id(elementId)))
-      .not.toBeVisible()
-      .withTimeout(10000);
-  }
-
-  /**
-   * @deprecated Use Assertion Class to accomplish this.
-   */
-  static async checkIfElementWithTextIsNotVisible(text) {
-    return await expect(element(by.text(text)).atIndex(0)).not.toBeVisible();
-  }
-
-  /**
-   * @deprecated Use Assertion Class to accomplish this.
-   */
-  static async checkIfElementNotToHaveText(elementId, text) {
-    await waitFor(element(by.id(elementId)))
-      .toBeVisible()
-      .withTimeout(10000);
-
-    return expect(element(by.id(elementId))).not.toHaveText(text);
-  }
-
-  /**
-   * @deprecated Use Assertion Class to accomplish this.
-   */
-  static async checkIfExists(elementId) {
-    await waitFor(element(by.id(elementId)))
-      .toBeVisible()
-      .withTimeout(10000);
-    return expect(element(by.id(elementId))).toExist();
-  }
-
-  /**
-   * @deprecated Use Assertion Class to accomplish this.
-   */
-  static async checkIfHasText(elementId, text) {
-    await waitFor(element(by.id(elementId)))
-      .toBeVisible()
-      .withTimeout(10000);
-
-    return expect(element(by.id(elementId))).toHaveText(text);
-  }
-
-  /**
-   * @deprecated Use Assertion Class to accomplish this.
-   */
-  static async checkIfElementWithTextIsVisible(text, index) {
-    return await waitFor(element(by.text(text)).atIndex(index || 0))
-      .toBeVisible()
-      .withTimeout(10000);
-  }
-
-  /**
-   * @deprecated Use Assertion Class to accomplish this.
-   */
-  static async checkIfElementByTextIsVisible(text, timeout = 25000) {
-    return await waitFor(element(by.text(text)))
-      .toBeVisible()
-      .withTimeout(timeout);
-  }
-
-  /**
-   * @deprecated Use Assertion Class to accomplish this.
-   */
-  static async checkIfElementHasString(elementID, text) {
-    return expect(element(by.id(elementID))).toString(text);
-  }
-
-  /**
-   * @deprecated Use Assertion Class to accomplish this.
-   */
-  static checkIfToggleIsOn(elementID) {
-    return expect(element(by.id(elementID))).toHaveToggleValue(true);
-  }
-
-  /**
-   * @deprecated Use Assertion Class to accomplish this.
-   */
-  static checkIfToggleIsOff(elementID) {
-    return expect(element(by.id(elementID))).toHaveToggleValue(false);
-  }
-
-  static relaunchApp() {
-    return this.launchApp({
-      newInstance: true,
-      launchArgs: {
-        detoxURLBlacklistRegex: Utilities.BlacklistURLs,
-      },
-    });
-  }
-
-  static delay(ms) {
-    return new Promise((resolve) => {
-      setTimeout(() => {
-        resolve();
-      }, ms);
-    });
-  } // Detox has no waits for webview elements visibility. Here is the custom one.
-
-  /**
-   * @deprecated Use Assertion Class to accomplish this.
-   */
-  static async waitForWebElementToBeVisibleById(elementId, timeout = 15000) {
-    const start = Date.now();
-    while (Date.now() - start < timeout) {
-      try {
-        await expect(web.element(by.web.id(elementId))).toExist(); // Element found
-        return;
-      } catch {
-        // Element not found yet: waiting for 200ms
-        await new Promise((resolve) => setTimeout(resolve, 200));
-      }
-    }
-    throw new Error('Element with ' + elementId + ' not found');
-  }
-  /**
-   * @deprecated Use Assertion Class to accomplish this.
-   */
-
-  static async retry(maxAttempts, testLogic) {
-    for (let attempt = 1; attempt <= maxAttempts; attempt++) {
-      try {
-        await testLogic();
-        return;
-      } catch (error) {
-        if (attempt === maxAttempts) {
-          throw error;
-        } else {
-          logger.error('Test attempt failed', {
-            attempt,
-            error,
-          });
-        }
-      }
-    }
-  }
-
-  static async reverseServerPort() {
-    // Port forwarding is now handled automatically in FixtureUtils.startResourceWithRetry()
-    // and startMultiInstanceResourceWithRetry() when ports are allocated.
-    // This function is kept for backwards compatibility with existing tests.
-    // will be removed in future.
-    logger.debug(
-      'reverseServerPort() called but port forwarding is now automatic',
-    );
-    return;
-  }
-
-  static async terminateApp() {
-    return device.terminateApp();
-  }
-
-  static async launchApp(launchOptions) {
-    const config = await resolveConfig();
-    const platform = device.getPlatform();
-    // Use debug launch for configs explicitly named 'debug' (original behavior)
-    // AND for any non-CI config (e.g. ios.sim.main which uses ios.debug app locally).
-    // CI configs (*.ci) use release apps and the normal recovery-based launch.
-    if (
-      config.configurationName.endsWith('debug') ||
-      !config.configurationName.endsWith('.ci')
-    ) {
-      return this.launchAppForDebugBuild(platform, launchOptions);
-    }
-
-    return this.launchAppWithRecovery(launchOptions);
-  }
-
-  static async launchAppForDebugBuild(platform, launchOptions) {
-    const deepLinkUrl = this.getDeepLinkUrl(
-      this.getDevLauncherPackagerUrl(platform),
-    );
-
-    if (platform === 'ios') {
-      await this.launchAppWithRecovery(launchOptions);
-      return device.openURL({
-        url: deepLinkUrl,
-      });
-    }
-
-    return this.launchAppWithRecovery({
-      url: deepLinkUrl,
-      ...launchOptions,
-    });
-  }
-
-  static getAdbDeviceFlag() {
-    return device?.id ? `-s ${device.id}` : '';
-  }
-
-  static async logAndroidReversePorts(context) {
-    if (device.getPlatform() !== 'android') {
-      return;
-    }
-
-    try {
-      const command = `adb ${this.getAdbDeviceFlag()} reverse --list`;
-      const { stdout } = await execAsync(command);
-      logger.debug(
-        `[launch diagnostics] ${context} | adb reverse --list:\n${
-          stdout?.trim() || '(empty)'
-        }`,
-      );
-    } catch (error) {
-      logger.warn(
-        `[launch diagnostics] ${context} | Failed to list adb reverse ports`,
-        error,
-      );
-    }
-  }
-
-  static isAndroidReverseAddressInUseError(error) {
-    const message = error instanceof Error ? error.message : String(error);
-    return (
-      device.getPlatform() === 'android' &&
-      message.includes('adb') &&
-      message.includes('reverse tcp:') &&
-      message.includes('Address already in use')
-    );
-  }
-
-  static extractConflictingReversePort(error) {
-    const message = error instanceof Error ? error.message : String(error);
-    const match = message.match(/reverse tcp:(\d+) tcp:\1/);
-    return match?.[1];
-  }
-
-  static async removeAndroidReversePort(port) {
-    try {
-      const command = `adb ${this.getAdbDeviceFlag()} reverse --remove tcp:${port}`;
-      await execAsync(command);
-      logger.warn(
-        `[launch recovery] Removed conflicting adb reverse tcp:${port}. Retrying app launch once.`,
-      );
-    } catch (removeError) {
-      const msg =
-        removeError instanceof Error
-          ? removeError.message
-          : String(removeError);
-      if (msg.includes('not found')) {
-        logger.warn(
-          `[launch recovery] Port tcp:${port} not in adb reverse table (likely TIME_WAIT or OS-level bind). Will still retry.`,
-        );
-      } else {
-        logger.error(
-          `[launch recovery] Unexpected error removing tcp:${port}: ${msg}`,
-        );
-      }
-    }
-  }
-
-  static async diagnoseEmulatorPort(port) {
-    const deviceFlag = this.getAdbDeviceFlag();
-    try {
-      // Dump socket state inside the emulator to identify what holds the port
-      const hexPort = parseInt(port, 10)
-        .toString(16)
-        .toUpperCase()
-        .padStart(4, '0');
-      const { stdout: tcpState } = await execAsync(
-        `adb ${deviceFlag} shell cat /proc/net/tcp6 2>/dev/null || adb ${deviceFlag} shell cat /proc/net/tcp 2>/dev/null`,
-      );
-      // Filter for the specific port (column 2 = local_address, port is after the colon in hex)
-      const relevantLines = tcpState
-        .split('\n')
-        .filter((line) => line.includes(`:${hexPort}`));
-      if (relevantLines.length > 0) {
-        logger.warn(
-          `[launch diagnostics] Emulator socket state for port ${port} (hex ${hexPort}):\n${relevantLines.join('\n')}`,
-        );
-        // TCP states: 01=ESTABLISHED, 02=SYN_SENT, 06=TIME_WAIT, 0A=LISTEN
-        logger.warn(
-          `[launch diagnostics] TCP state reference: 01=ESTABLISHED, 06=TIME_WAIT, 0A=LISTEN`,
-        );
-      } else {
-        logger.warn(
-          `[launch diagnostics] No socket entries found for port ${port} inside emulator`,
-        );
-      }
-    } catch (diagError) {
-      logger.debug(
-        `[launch diagnostics] Could not read emulator socket state: ${diagError.message}`,
-      );
-    }
-  }
-
-  static async launchAppWithRecovery(launchOptions) {
-    await this.logAndroidReversePorts('before device.launchApp');
-
-    try {
-      return await device.launchApp(launchOptions);
-    } catch (error) {
-      if (!this.isAndroidReverseAddressInUseError(error)) {
-        throw error;
-      }
-
-      const conflictingPort = this.extractConflictingReversePort(error);
-      if (!conflictingPort) {
-        throw error;
-      }
-
-      logger.warn(
-        `[launch recovery] Detected Detox adb reverse port collision on tcp:${conflictingPort}.`,
-      );
-      await this.logAndroidReversePorts(
-        'launch failure before reverse cleanup',
-      );
-      await this.diagnoseEmulatorPort(conflictingPort);
-      await this.removeAndroidReversePort(conflictingPort);
-      await this.logAndroidReversePorts('after reverse cleanup before retry');
-
-      return device.launchApp(launchOptions);
-    }
-  }
-
-  static getDeepLinkUrl(url) {
-    return `expo-metamask://expo-development-client/?url=${encodeURIComponent(
-      url,
-    )}`;
-  }
-
-  static getDevLauncherPackagerUrl(platform) {
-    const port =
-      process.env.METRO_PORT_E2E || process.env.WATCHER_PORT || '8081';
-    return `http://localhost:${port}/index.bundle?platform=${platform}&dev=true&minify=false&disableOnboarding=1`;
-  }
-}
