+++ tests/page-objects/wallet/WalletView.ts
@@ -5,15 +5,9 @@ import {
 import { EARN_TEST_IDS } from '../../../app/components/UI/Earn/constants/testIds';
 import { CashGetMusdEmptyStateSelectors } from '../../../app/components/Views/Homepage/Sections/Cash/CashGetMusdEmptyState.testIds';
 import { SECONDARY_BALANCE_BUTTON_TEST_ID } from '../../../app/components/UI/AssetElement/index.constants';
-import {
-  PredictPositionsHeaderSelectorsIDs,
-  PredictClaimConfirmationSelectorsIDs,
-  PredictMarketDetailsSelectorsIDs,
-} from '../../../app/components/UI/Predict/Predict.testIds';
 import Gestures from '../../framework/Gestures';
 import UnifiedGestures from '../../framework/UnifiedGestures';
 import Matchers from '../../framework/Matchers';
-import TestHelpers from '../../helpers.js';
 import Assertions from '../../framework/Assertions';
 import PlaywrightAssertions from '../../framework/PlaywrightAssertions';
 import Utilities from '../../framework/Utilities';
@@ -25,13 +19,13 @@ import {
 } from '../../framework/EncapsulatedElement';
 import { encapsulatedAction } from '../../framework/encapsulatedAction';
 import PlaywrightMatchers from '../../framework/PlaywrightMatchers';
-import { PlatformDetector } from '../../framework/PlatformLocator';
-import { FrameworkDetector } from '../../framework/FrameworkDetector';
 import PlaywrightGestures from '../../framework/PlaywrightGestures';
+import { PlatformDetector } from '../../framework/PlatformLocator';
 import { getAssetTestId } from '../../selectors/Wallet/WalletView.selectors';
-import { resolveE2EWaitTimeoutMs } from '../../framework/Constants';
+import WalletHomeScroll from './WalletHomeScroll';
+import { WalletHomeSections as WalletHomeSectionsBase } from './WalletHomeSections';
 
-class WalletView {
+class WalletView extends WalletHomeSectionsBase {
   get container(): EncapsulatedElementType {
     return Matchers.getElementByID(WalletViewSelectorsIDs.WALLET_CONTAINER);
   }
@@ -59,259 +53,12 @@ class WalletView {
     });
   }
 
-  /** Cross-framework scroll container for the wallet homepage ScrollView. */
   get walletScrollContainer(): string {
-    return WalletViewSelectorsIDs.WALLET_SCROLL_VIEW;
+    return WalletHomeScroll.walletScrollContainer;
   }
 
-  /** Wallet ScrollView as element (for gestures like swipe). */
   get walletScrollView(): EncapsulatedElementType {
-    return Matchers.getElementByID(WalletViewSelectorsIDs.WALLET_SCROLL_VIEW);
-  }
-
-  private isAndroidAppium(): boolean {
-    return FrameworkDetector.isAppium() && !PlatformDetector.isIOS();
-  }
-
-  /**
-   * Map Detox scroll direction to swipe args for `wallet-scroll-view` (iOS).
-   */
-  private mapWalletHomeScrollToSwipe(
-    scrollDirection: 'up' | 'down',
-  ): 'up' | 'down' {
-    return scrollDirection === 'down' ? 'up' : 'down';
-  }
-
-  /**
-   * Scroll the wallet homepage by swiping inside `wallet-scroll-view`.
-   * `scrollDirection` matches Detox semantics (`down` reveals content below).
-   */
-  private async scrollWalletHome(
-    scrollDirection: 'up' | 'down',
-    percent = 0.45,
-  ): Promise<void> {
-    if (this.isAndroidAppium()) {
-      await this.scrollWalletHomeAndroid(scrollDirection, percent);
-      return;
-    }
-
-    const swipeDirection = this.mapWalletHomeScrollToSwipe(scrollDirection);
-    await UnifiedGestures.swipe(this.walletScrollView, swipeDirection, {
-      percentage: percent,
-      description: `Scroll wallet homepage ${scrollDirection}`,
-    });
-  }
-
-  /**
-   * Android wallet homepage: use bounded W3C swipes inside the scroll view.
-   * `mobile:scrollGesture` with `up` triggers pull-to-refresh; starting the
-   * gesture near the top edge is unreliable even with `direction: down`.
-   */
-  private async scrollWalletHomeAndroid(
-    scrollDirection: 'up' | 'down',
-    percent = 0.45,
-  ): Promise<void> {
-    const container = await asPlaywrightElement(this.walletScrollView);
-    const location = await container.unwrap().getLocation();
-    const size = await container.unwrap().getSize();
-    const centerX = Math.floor(location.x + size.width / 2);
-    const travel = Math.floor(
-      size.height * Math.min(Math.max(percent, 0.1), 0.9),
-    );
-    const fingerDirection = scrollDirection === 'down' ? 'up' : 'down';
-
-    const fromY =
-      fingerDirection === 'up'
-        ? location.y + Math.floor(size.height * 0.75)
-        : location.y + Math.floor(size.height * 0.35);
-    const toY = fingerDirection === 'up' ? fromY - travel : fromY + travel;
-
-    await PlaywrightGestures.swipe({
-      scrollParams: { direction: fingerDirection },
-      percent,
-      duration: 600,
-      from: { x: centerX, y: fromY },
-      to: { x: centerX, y: toY },
-    });
-  }
-
-  /** Scroll until `target` is visible on the wallet homepage. */
-  private async scrollWalletHomeToElement(
-    target: EncapsulatedElementType,
-    description: string,
-    direction: 'up' | 'down' = 'down',
-    maxAttempts = 16,
-  ): Promise<void> {
-    // Android: scrollIntoView inside the scroll view works reliably.
-    // iOS: scrollIntoView gets stuck on the homepage; use bounded swipe loops instead.
-    if (this.isAndroidAppium()) {
-      await Assertions.expectElementToBeVisible(this.walletScrollView, {
-        timeout: resolveE2EWaitTimeoutMs(10_000),
-        description: `wallet-scroll-view for ${description}`,
-      });
-      const scrollView = await asPlaywrightElement(this.walletScrollView);
-      const element = await asPlaywrightElement(target);
-      await PlaywrightGestures.scrollIntoView(element, {
-        scrollableElement: scrollView,
-        scrollParams: {
-          direction: direction === 'down' ? 'up' : 'down',
-        },
-        maxScrolls: maxAttempts,
-      });
-      await Assertions.expectElementToBeVisible(target, {
-        timeout: 5_000,
-        description,
-      });
-      return;
-    }
-
-    if (FrameworkDetector.isAppium()) {
-      for (let attempt = 0; attempt < maxAttempts; attempt += 1) {
-        try {
-          await Assertions.expectElementToBeVisible(target, {
-            timeout: 1_500,
-            description,
-          });
-          return;
-        } catch {
-          await this.scrollWalletHome(direction, 0.5);
-        }
-      }
-
-      await Assertions.expectElementToBeVisible(target, {
-        timeout: 5_000,
-        description,
-      });
-      return;
-    }
-
-    for (let attempt = 0; attempt < maxAttempts; attempt += 1) {
-      try {
-        await Assertions.expectElementToBeVisible(target, {
-          timeout: 1_500,
-          description,
-        });
-        return;
-      } catch {
-        await this.scrollWalletHome(direction, 0.5);
-      }
-    }
-
-    await Assertions.expectElementToBeVisible(target, {
-      timeout: 5_000,
-      description,
-    });
-  }
-
-  private async tapIfAlreadyVisible(
-    target: DetoxElement | EncapsulatedElementType,
-    description: string,
-    options: { tapTimeout?: number } = {},
-  ): Promise<boolean> {
-    if (!FrameworkDetector.isAppium()) {
-      return false;
-    }
-
-    const { tapTimeout = 30_000 } = options;
-
-    try {
-      await PlaywrightAssertions.expectElementToBeVisible(
-        asPlaywrightElement(target),
-        { timeout: 2000, description },
-      );
-      await Gestures.waitAndTap(target, {
-        elemDescription: description,
-        timeout: tapTimeout,
-      });
-      return true;
-    } catch {
-      return false;
-    }
-  }
-
-  /**
-   * Progressive scroll for homepage sections:
-   * try tap -> small scroll down -> retry, until the section is tappable.
-   * @param options.scrollAmount - Pixels to scroll per step.
-   * @param options.overshootSwipe - After scroll, perform a small swipe to move the section away
-   * from the tab bar (e.g. direction 'up' = one more scroll down = section moves higher on screen).
-   */
-  private async scrollAndTapSection(
-    target: DetoxElement | EncapsulatedElementType,
-    description: string,
-    direction: 'up' | 'down' = 'down',
-    options: {
-      scrollAmount?: number;
-      overshootSwipe?: { direction: 'up' | 'down'; percentage?: number };
-      timeout?: number;
-      /** Appium/Detox tap wait after scroll. Defaults to 30s. */
-      tapTimeout?: number;
-    } = {},
-  ): Promise<void> {
-    const {
-      scrollAmount = 200,
-      overshootSwipe,
-      timeout = 15_000,
-      tapTimeout = 30_000,
-    } = options;
-
-    await encapsulatedAction({
-      detox: async () => {
-        await Gestures.scrollToElement(target, this.walletScrollContainer, {
-          direction,
-          scrollAmount,
-          timeout,
-          elemDescription: `Scroll to ${description}`,
-        });
-        if (overshootSwipe) {
-          await Gestures.swipe(
-            this.walletScrollView,
-            overshootSwipe.direction,
-            {
-              percentage: overshootSwipe.percentage ?? 0.15,
-              speed: 'slow',
-              elemDescription: `Overshoot swipe for ${description}`,
-            },
-          );
-        }
-        await Gestures.waitAndTap(target, {
-          elemDescription: description,
-          timeout: tapTimeout,
-        });
-      },
-      appium: async () => {
-        await this.scrollWalletHomeToElement(
-          target as EncapsulatedElementType,
-          description,
-          direction,
-          Math.max(8, Math.ceil(timeout / 2_000)),
-        );
-        if (overshootSwipe) {
-          const overshootScrollDirection =
-            overshootSwipe.direction === 'up' ? 'down' : 'up';
-          if (this.isAndroidAppium()) {
-            await this.scrollWalletHomeAndroid(
-              overshootScrollDirection,
-              overshootSwipe.percentage ?? 0.15,
-            );
-          } else {
-            await Gestures.swipe(
-              this.walletScrollView,
-              overshootSwipe.direction,
-              {
-                percentage: overshootSwipe.percentage ?? 0.15,
-                speed: 'slow',
-                elemDescription: `Overshoot swipe for ${description}`,
-              },
-            );
-          }
-        }
-        await PlaywrightGestures.waitAndTap(
-          await asPlaywrightElement(target as EncapsulatedElementType),
-          { timeout: tapTimeout },
-        );
-      },
-    });
+    return WalletHomeScroll.walletScrollView;
   }
 
   get earnButton(): EncapsulatedElementType {
@@ -612,7 +359,12 @@ class WalletView {
   }
 
   async tapNetworksButtonOnNavBar(): Promise<void> {
-    await TestHelpers.tap(WalletViewSelectorsIDs.NAVBAR_NETWORK_BUTTON);
+    await Gestures.waitAndTap(
+      Matchers.getElementByID(WalletViewSelectorsIDs.NAVBAR_NETWORK_BUTTON),
+      {
+        elemDescription: 'Navbar Network Button',
+      },
+    );
   }
 
   async tapNavbarCardButton(): Promise<void> {
@@ -704,450 +456,6 @@ class WalletView {
       speed: 'slow',
       percentage: 0.7,
       elemDescription: 'Swipe Carousel',
-    });
-  }
-
-  get defiTab(): EncapsulatedElementType {
-    return Matchers.getElementByText(WalletViewSelectorsText.DEFI_TAB);
-  }
-
-  get defiNetworkFilter(): EncapsulatedElementType {
-    return Matchers.getElementByID(
-      WalletViewSelectorsIDs.DEFI_POSITIONS_NETWORK_FILTER,
-    );
-  }
-
-  get defiTabContainer(): EncapsulatedElementType {
-    return Matchers.getElementByID(
-      WalletViewSelectorsIDs.DEFI_POSITIONS_CONTAINER,
-    );
-  }
-  get claimButton(): EncapsulatedElementType {
-    return Matchers.getElementByID(
-      PredictPositionsHeaderSelectorsIDs.CLAIM_BUTTON,
-    );
-  }
-  get predictClaimConfirmButton(): EncapsulatedElementType {
-    return Matchers.getElementByID(
-      PredictClaimConfirmationSelectorsIDs.CLAIM_CONFIRM_BUTTON,
-    );
-  }
-
-  get availableBalanceLabel(): EncapsulatedElementType {
-    return Matchers.getElementByText(WalletViewSelectorsText.AVAILABLE_BALANCE);
-  }
-
-  get defiPositionsNew(): EncapsulatedElementType {
-    return Matchers.getElementByText(WalletViewSelectorsText.DEFI_SECTION);
-  }
-
-  /** Perpetuals section title button on the homepage. */
-  get perpsSectionHeader(): EncapsulatedElementType {
-    return Matchers.getElementByID(
-      WalletViewSelectorsIDs.HOMEPAGE_SECTION_TITLE('perps'),
-    );
-  }
-
-  /** Predictions section title button on the homepage. */
-  get predictionsSectionHeader(): EncapsulatedElementType {
-    return Matchers.getElementByID(
-      WalletViewSelectorsIDs.HOMEPAGE_SECTION_TITLE('predictions'),
-    );
-  }
-
-  get getMoneySection(): EncapsulatedElementType {
-    return PlaywrightMatchers.getElementById('homepage-section-title-cash', {
-      exact: false,
-    });
-  }
-  /** Tokens section header on the homepage. */
-  get tokensSectionHeader(): EncapsulatedElementType {
-    return Matchers.getElementByText(WalletViewSelectorsText.TOKENS_SECTION);
-  }
-
-  get tokensSection(): EncapsulatedElementType {
-    return encapsulated({
-      detox: () =>
-        Matchers.getElementByID(
-          WalletViewSelectorsIDs.TOKENS_SECTION_CONTAINER,
-        ),
-      appium: {
-        android: () =>
-          PlaywrightMatchers.getElementById(
-            WalletViewSelectorsIDs.HOMEPAGE_SECTION_TITLE('tokens'),
-          ),
-        ios: () =>
-          PlaywrightMatchers.getElementByAccessibilityId(
-            `${WalletViewSelectorsIDs.HOMEPAGE_SECTION_TITLE('tokens')}`,
-          ),
-      },
-    });
-  }
-
-  /** NFTs section header on the homepage. */
-  get nftsSectionHeader(): EncapsulatedElementType {
-    return Matchers.getElementByText(WalletViewSelectorsText.NFTS_SECTION);
-  }
-
-  async tapOnNewTokensSection(): Promise<void> {
-    await encapsulatedAction({
-      detox: async () => {
-        await Gestures.waitAndTap(this.tokensSectionHeader, {
-          checkStability: true,
-          elemDescription: 'New Tokens Section',
-        });
-      },
-      appium: async () => {
-        await this.scrollAndTapTokensSection();
-      },
-    });
-  }
-
-  /**
-   * Scrolls the wallet homepage until the Tokens section header is tappable.
-   * On long homepages the header can sit below the fold near the tab bar; a plain
-   * tap then hits the tab bar instead of navigating to Tokens Full View.
-   */
-  async scrollAndTapTokensSection(
-    direction: 'up' | 'down' = 'down',
-  ): Promise<void> {
-    if (
-      await this.tapIfAlreadyVisible(this.tokensSectionHeader, 'Tokens section')
-    ) {
-      return;
-    }
-
-    const getScrollOptions = (scrollDirection: 'up' | 'down') => ({
-      overshootSwipe: {
-        direction:
-          scrollDirection === 'down' ? ('up' as const) : ('down' as const),
-        percentage: 0.15,
-      },
-      timeout: 60_000,
-    });
-
-    try {
-      await this.scrollAndTapSection(
-        this.tokensSectionHeader,
-        'Tokens section',
-        direction,
-        getScrollOptions(direction),
-      );
-    } catch {
-      const fallbackDirection = direction === 'down' ? 'up' : 'down';
-      await this.scrollAndTapSection(
-        this.tokensSectionHeader,
-        'Tokens section',
-        fallbackDirection,
-        getScrollOptions(fallbackDirection),
-      );
-    }
-  }
-
-  async tapOnTokensSection(): Promise<void> {
-    await encapsulatedAction({
-      detox: async () => {
-        await Gestures.waitAndTap(this.tokensSectionHeader, {
-          checkStability: true,
-          elemDescription: 'Tokens Section',
-        });
-      },
-      appium: async () => {
-        const elem = await asPlaywrightElement(this.tokensSection);
-        await PlaywrightGestures.waitForElementStable(elem);
-
-        // Re-fetch to avoid stale reference after stability wait
-        const freshElem = await asPlaywrightElement(this.tokensSection);
-        await freshElem.unwrap().click();
-      },
-    });
-  }
-
-  async tapOnDeFiTab(): Promise<void> {
-    await Gestures.waitAndTap(this.defiTab, {
-      elemDescription: 'DeFi Tab',
-    });
-  }
-
-  async tapOnDeFiNetworksFilter(): Promise<void> {
-    await Gestures.waitAndTap(this.defiNetworkFilter, {
-      elemDescription: 'DeFi Networks Filter',
-    });
-  }
-
-  async scrollAndTapDefiSection(): Promise<void> {
-    await this.scrollAndTapSection(this.defiPositionsNew, 'DeFi section');
-  }
-
-  async scrollAndTapPerpsSection(): Promise<void> {
-    try {
-      await this.scrollAndTapSection(
-        this.perpsSectionHeader,
-        'Perpetuals section',
-      );
-    } catch {
-      await this.scrollAndTapSection(
-        this.perpsSectionHeader,
-        'Perpetuals section',
-        'up',
-      );
-    }
-  }
-
-  /**
-   * Scrolls to the Predictions section and taps it. After scroll, does a small overshoot swipe
-   * so the section sits higher on screen and the tap does not hit the main menu "+" button.
-   */
-  async scrollAndTapPredictionsSection(
-    direction: 'up' | 'down' = 'down',
-    options: {
-      overshootSwipe?: { direction: 'up' | 'down'; percentage?: number };
-    } = {},
-  ): Promise<void> {
-    if (
-      await this.tapIfAlreadyVisible(
-        this.predictionsSectionHeader,
-        'Predictions section',
-      )
-    ) {
-      return;
-    }
-
-    const fallbackDirection = direction === 'down' ? 'up' : 'down';
-
-    await this.tryScrollDirections(
-      (scrollDirection) =>
-        this.scrollAndTapSection(
-          this.predictionsSectionHeader,
-          'Predictions section',
-          scrollDirection,
-          {
-            overshootSwipe: options.overshootSwipe ?? {
-              direction:
-                scrollDirection === 'down'
-                  ? ('up' as const)
-                  : ('down' as const),
-              percentage: 0.15,
-            },
-            timeout: 60_000,
-          },
-        ),
-      [direction, fallbackDirection],
-    );
-  }
-
-  private async scrollPredictionsSectionIntoView(
-    direction: 'up' | 'down' = 'down',
-    options: { maxAttempts?: number } = {},
-  ): Promise<void> {
-    const { maxAttempts = 24 } = options;
-
-    await encapsulatedAction({
-      detox: async () => {
-        await Gestures.scrollToElement(
-          this.predictionsSectionHeader,
-          this.walletScrollContainer,
-          {
-            direction,
-            scrollAmount: 250,
-            timeout: 60_000,
-            elemDescription: 'Scroll to Predictions section',
-          },
-        );
-      },
-      appium: async () => {
-        await this.scrollWalletHomeToElement(
-          this.predictionsSectionHeader,
-          'Predictions section',
-          direction,
-          maxAttempts,
-        );
-      },
-    });
-  }
-
-  /**
-   * Tries scroll actions down then up (or a custom order) without nested try/catch
-   * at call sites. Rethrows the last failure if every direction fails.
-   */
-  private async tryScrollDirections(
-    action: (direction: 'up' | 'down') => Promise<void>,
-    directions: readonly ('up' | 'down')[] = ['down', 'up'],
-  ): Promise<void> {
-    let lastError: unknown;
-    for (const direction of directions) {
-      try {
-        await action(direction);
-        return;
-      } catch (error) {
-        lastError = error;
-      }
-    }
-    throw lastError;
-  }
-
-  async scrollAndTapPredictionsPosition(
-    positionName: string,
-    positionId?: string,
-  ): Promise<void> {
-    const target = positionId
-      ? Matchers.getElementByID(`predict-position-row-${positionId}`)
-      : Matchers.getElementByText(positionName);
-
-    await encapsulatedAction({
-      detox: async () => {
-        if (
-          await this.tapIfAlreadyVisible(
-            target,
-            `Predictions Position: ${positionName}`,
-          )
-        ) {
-          return;
-        }
-
-        await this.tryScrollDirections((direction) =>
-          this.scrollPredictionsSectionIntoView(direction),
-        );
-
-        if (
-          await this.tapIfAlreadyVisible(
-            target,
-            `Predictions Position: ${positionName}`,
-          )
-        ) {
-          return;
-        }
-
-        await this.tryScrollDirections((direction) =>
-          this.scrollAndTapSection(
-            target,
-            `Predictions Position: ${positionName}`,
-            direction,
-            { timeout: 60_000 },
-          ),
-        );
-      },
-      appium: async () => {
-        const description = `Predictions Position: ${positionName}`;
-        const marketDetailsScreen = Matchers.getElementByID(
-          PredictMarketDetailsSelectorsIDs.SCREEN,
-        );
-        // Keep one attempt well under half of the outer budget so
-        // tryScrollDirections (down+up) can fail and still leave a second try.
-        const scrollAndTapRetryTimeoutMs = 90_000;
-        const intoViewMaxAttempts = 8;
-        const scrollAndTapPerDirectionMs = 15_000;
-        const tapTimeoutMs = 10_000;
-        const predictNavigationTimeoutMs = resolveE2EWaitTimeoutMs(15_000);
-
-        const assertMarketDetailsOpened = async (): Promise<void> => {
-          await Assertions.expectElementToBeVisible(marketDetailsScreen, {
-            timeout: predictNavigationTimeoutMs,
-            description: 'Predict market details screen after position tap',
-          });
-        };
-
-        // Align with Detox: tap-if-visible → section into view → tap →
-        // scrollAndTap with overshoot (down/up). Do not require the row to be
-        // visible before scrolling (off-screen rows like Blue Jays).
-        await Utilities.executeWithRetry(
-          async () => {
-            if (
-              await this.tapIfAlreadyVisible(target, description, {
-                tapTimeout: tapTimeoutMs,
-              })
-            ) {
-              await assertMarketDetailsOpened();
-              return;
-            }
-
-            await this.tryScrollDirections((direction) =>
-              this.scrollPredictionsSectionIntoView(direction, {
-                maxAttempts: intoViewMaxAttempts,
-              }),
-            );
-
-            if (
-              await this.tapIfAlreadyVisible(target, description, {
-                tapTimeout: tapTimeoutMs,
-              })
-            ) {
-              await assertMarketDetailsOpened();
-              return;
-            }
-
-            await this.tryScrollDirections((direction) =>
-              this.scrollAndTapSection(target, description, direction, {
-                timeout: scrollAndTapPerDirectionMs,
-                tapTimeout: tapTimeoutMs,
-                overshootSwipe: {
-                  direction: direction === 'down' ? 'up' : 'down',
-                  percentage: 0.15,
-                },
-              }),
-            );
-
-            await assertMarketDetailsOpened();
-          },
-          {
-            timeout: scrollAndTapRetryTimeoutMs,
-            description: `Scroll and tap ${description}`,
-          },
-        );
-      },
-    });
-  }
-
-  async scrollAndTapNftsSection(): Promise<void> {
-    await this.scrollAndTapSection(this.nftsSectionHeader, 'NFTs section');
-  }
-
-  async tapOnAvailableBalance(): Promise<void> {
-    await Gestures.waitAndTap(this.availableBalanceLabel, {
-      elemDescription: 'tap available balance to expand balance card',
-    });
-  }
-
-  async tapClaimButton(): Promise<void> {
-    await encapsulatedAction({
-      detox: async () => {
-        await Gestures.scrollToElement(
-          this.claimButton,
-          this.walletScrollContainer,
-          {
-            direction: 'down',
-            scrollAmount: 200,
-            elemDescription: 'Scroll to Claim Button',
-          },
-        );
-        await Gestures.waitAndTap(this.claimButton, {
-          elemDescription: 'Claim Button',
-        });
-      },
-      appium: async () => {
-        await Utilities.executeWithRetry(
-          async () => {
-            await this.scrollWalletHomeToElement(
-              this.claimButton,
-              'Claim button on wallet homepage',
-            );
-            await UnifiedGestures.waitAndTap(this.claimButton, {
-              description: 'Claim Button',
-              timeout: 30_000,
-            });
-          },
-          {
-            timeout: 90_000,
-            description: 'Tap claim button on wallet homepage',
-          },
-        );
-      },
-    });
-  }
-
-  async tapClaimConfirmButton(): Promise<void> {
-    await Gestures.waitAndTap(this.predictClaimConfirmButton, {
-      elemDescription: 'Claim confirm button',
     });
   }
